Project_2_MPesa_Airflow_Pipeline/requirements.txt
apache-airflow>=2.6.0
pandas>=2.0.0
numpy>=1.24.0
faker>=18.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pytest>=7.0.0
